refactor: Controller 의존관계 변경
[AS-IS] Controller -> Service -> Service (Model 활용 X)
[TO-BE] Controller -> Service, View (Model 활용 O)
